pyproject.toml
[project]
name = "dida-mcp"
version = "1.0.0"
description = "滴答清单 (Dida365/TickTick) MCP Server - 让 AI 助手控制你的滴答清单"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "dida-skill contributors"},
]
keywords = ["mcp", "dida365", "ticktick", "todo", "ai", "openclaw"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dida-mcp = "dida_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/Martinqi826/dida-skill"
Repository = "https://github.com/Martinqi826/dida-skill"
Issues = "https://github.com/Martinqi826/dida-skill/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
